Disable autocast on crf loss calculations

+++ models_mage_2dec.py
@@ -915,7 +915,7 @@ class MaskedGenerativeEncoderViT(nn.Module):
                 attn_logits = F.interpolate(attn_logits, size=mask_crf.shape[-1], mode='bilinear')
                 attn_interpolate = F.interpolate(attn_transpose, size=mask_crf.shape[-1], mode='bilinear')
                 attn_onehot = torch.nn.functional.one_hot(attn_interpolate.argmax(1), num_classes=self.slot_attention.num_slots).permute(0,3,1,2)
-                mask_crf_onehot = torch.nn.functional.one_hot(mask_crf, num_classes=self.slot_attention.num_slots).permute(0,3,1,2).to(dtype=torch.float16)
+                mask_crf_onehot = torch.nn.functional.one_hot(mask_crf, num_classes=self.slot_attention.num_slots).permute(0,3,1,2).to(dtype=torch.float32)
 
                 permutation_indices, _ = att_matching(attn_onehot, mask_crf_onehot)
